requirements.txt
accelerate==1.8.1
annotated-types==0.7.0
anyio==4.9.0
av==15.0.0
bitsandbytes==0.42.0
certifi==2022.12.7
charset-normalizer==2.1.1
decorator==5.2.1
distro==1.9.0
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.13.1
flash-attn==2.5.3
fsspec==2024.6.1
fvcore==0.1.5.post20221221
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
iopath==0.1.10
Jinja2==3.1.3
MarkupSafe==2.1.5
moviepy==2.2.1
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.4
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
openai==1.14.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.1
parameterized==0.9.0
pillow==11.0.0
portalocker==3.2.0
proglog==0.1.12
protobuf==6.31.1
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorchvideo==0.1.5
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.1
safetensors==0.5.3
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.3
tabulate==0.9.0
termcolor==3.1.0
tokenizers==0.21.2
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchvision==0.17.0+cu118
tqdm==4.67.1
transformers==4.53.2
transformers-stream-generator==0.0.4
triton==2.2.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.13
yacs==0.1.8
